+++ test/integration/user_registration_flows.rb
@@ -11,7 +11,7 @@ class UserRegistrationFlowsTest < ActionDispatch::IntegrationTest
 		DatabaseCleaner.clean
 	end
 	test "user registers by Facebook two step procedure" do
-    user_fields = 'first_name,last_name,email,gender,location'
+    user_fields = 'first_name,last_name,gender,location'
     Facebook.any_instance.stubs(:user_fields).returns(user_fields)
     # First step
     get '/auth', { provider: @provider, auth_token: @auth_token }
@@ -20,9 +20,9 @@ class UserRegistrationFlowsTest < ActionDispatch::IntegrationTest
     user = body[:data][:user]
     assert_equal 200, body[:status]
     assert_equal 102, body[:code], "require additional information to complete registration" 
-    assert_equal "can't be blank", body[:data][:errors][:age].join
+    assert_equal "can't be blank", body[:data][:errors][:email].join
     # Second step
-    user[:age] = 21
+    user[:email] = 'user@test.com'
     post '/users', { provider: @provider, auth_token: @auth_token, user: user }
     assert_response 201
     body = JSON.parse(response.body).deep_symbolize_keys
@@ -30,14 +30,13 @@ class UserRegistrationFlowsTest < ActionDispatch::IntegrationTest
     assert_equal 101, body[:code], "successfully registered"
     assert_equal 'Yaroslav', body[:data][:user][:first_name]
     assert_equal 'Nychka', body[:data][:user][:last_name]
-    assert_equal 'nychka08@yandex.ru', body[:data][:user][:email]
-    assert_equal 21, body[:data][:user][:age]
+    assert_equal 'user@test.com', body[:data][:user][:email]
     assert_not_nil body[:data][:user][:access_token]
   end
   test "user registers by social network two step procedure" do
     extra = { bdate: nil, socials: {} }
     extra[:socials][@provider.to_sym] = Settings[@provider]['site']
-    @data[:body] = attributes_for(:social_user, age: nil, url: Settings[@provider]['site']).merge(extra)
+    @data[:body] = attributes_for(:social_user, email: nil, url: Settings[@provider]['site']).merge(extra)
     Facebook.any_instance.stubs(:get_user_info).returns(@data)
     # First step
     get '/auth', { provider: @provider, auth_token: @auth_token }
@@ -47,15 +46,15 @@ class UserRegistrationFlowsTest < ActionDispatch::IntegrationTest
     assert_equal 102, body[:code], "require additional information to complete registration" 
     @data[:body].delete(:url)
     assert_equal @data[:body], body[:data][:user]
-    assert_equal "can't be blank", body[:data][:errors][:age].join
+    assert_equal "can't be blank", body[:data][:errors][:email].join
     # Second step
-    @data[:body][:age] = 22
+    @data[:body][:email] = 'test@user.com'
     post '/users', { provider: @provider, auth_token: @auth_token, user: @data[:body] }
     assert_response 201
     body = JSON.parse(response.body).deep_symbolize_keys
     assert_equal 201, body[:status]
     assert_equal 101, body[:code], "successfully registered"
-    assert_equal 22, body[:data][:user][:age]
+    assert_equal 'test@user.com', body[:data][:user][:email]
   end
 	test "user connects Vkontakte two step procedure" do
     email = 'test@mail.com'